Move InvalidStreamIdException to the correct package
